Added support and test cases for live camera and video input.
